+++ src/main/java/wooteco/subway/ui/dto/request/SectionRequest.java
@@ -1,30 +1,18 @@
-package wooteco.subway.service.dto.line;
+package wooteco.subway.ui.dto.request;
 
-public class LineRequest {
+public class SectionRequest {
 
-    private String name;
-    private String color;
     private Long upStationId;
     private Long downStationId;
     private int distance;
 
-    public LineRequest() {
+    public SectionRequest() {
     }
 
-    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
-        this.name = name;
-        this.color = color;
+    public SectionRequest(Long upStationId, Long downStationId, int distance) {
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
     }
 
     public Long getUpStationId() {
