+++ src/main/java/wooteco/subway/ui/dto/response/RouteResponse.java
@@ -1,7 +1,6 @@
-package wooteco.subway.service.dto;
+package wooteco.subway.ui.dto.response;
 
 import java.util.List;
-import wooteco.subway.service.dto.station.StationResponse;
 
 public class RouteResponse {
